requirements.txt
fastapi==0.128.8
requests==2.32.5 
uvicorn[standard]
python-dotenv==1.2.1
joblib=1.5.3
scikit-learn=1.6.1
pandas=2.3.3
